requirements.txt
torch
torchvision
torchaudio
scikit-learn
matplotlib
tqdm
pillow
